The implementation of the Context interface now has a Keyvals method.
This does not change the public interface, but it opens up some intriguing possibilities for unifying logging and error parameters.

For example:
```go
func doSomethingWith(ctx context.Context, userID string, documentID string) error {
    errors := errors.With("userID", userID, "documentID", documentID)
    log := log.FromContext(ctx).With(errors)

    // ... rest of function here ...
```

In this scenario, the errors context is created, so any errors created in the function will
have the userID and documentID attached. This function also happens to log messages,
and it wants the same context for logging. Because the errors context implements the
keyvalser interface, the log With method will accept it a s a list of key/value pairs.

This all pre-supposes that the logging framework is making use of package kv. Untried
as of yet, but it looks pretty promising.